refactor: create useEditable hook
